+++ lib/shared/open_ai/ai_function.dart
@@ -1,0 +1,16 @@
+class AiFunction {
+  final String name;
+  final String description;
+  final Map<String, dynamic> parameters;
+
+  AiFunction(
+      {required this.name,
+        required this.description,
+        required this.parameters});
+
+  Map<String, dynamic> toJson() => {
+    'name': name,
+    'description': description,
+    'parameters': parameters,
+  };
+}
