+++ lib/shared/open_ai/api_client.dart
@@ -1,6 +1,9 @@
 import 'dart:convert';
 import 'dart:developer' as developer;
 import 'package:http/http.dart' as http;
+import 'ai_function.dart';
+import 'content.dart';
+import 'message.dart';
 import 'request_body.dart';
 import '../../db/ingredient.dart';
 import '../../db/meal_plan_entry.dart';
