+++ lib/shared/open_ai/message.dart
@@ -1,0 +1,13 @@
+import 'content.dart';
+
+class Message {
+  final String role;
+  final List<Content> content;
+
+  Message({required this.role, required this.content});
+
+  Map<String, dynamic> toJson() => {
+    'role': role,
+    'content': content.map((c) => c.toJson()).toList(),
+  };
+}
